Bug fix: B50-ZK-630: Doublespinner exception with constraint http://www.zkoss.org/services/zk/tracker/?ZK-630
git-svn-id: https://svn.potix.com/svn/zk1/branches/5.0@18365 dd50bd9b-9913-0410-b9ba-e07a3075be86

+++ zul/src/org/zkoss/zul/Doublespinner.java
@@ -130,7 +130,7 @@ public class Doublespinner extends NumberInputElement implements
 
 		try {
 			double v = Double.parseDouble(val);
-			double divscale = vals[1] != null ? ((Double) vals[1]).doubleValue() : 0;
+			double divscale = vals[1] != null ? ((Integer) vals[1]).intValue() : 0;
 			while (v != 0 && --divscale >= 0)
 				v /= 10;
 			return new Double(v);
